added Nari Ultimate to razeraccessory driver
Moved Nari Ultimate to razeraccessory driver. First attempt to add compatibility for static effect based on @pieromarini 's script and research.

Changes don't work. Effect is shown in RazerGenie but doesn't change the color on the headphones. Since the headphones have a separate protocol from anything else, I believe a custom report had to be made for the driver as well as a custom Static Effect function.